refactor(frontend): consolidate SearchableSelect variants
